+++ bin/solarEclipse/downloadBesselianElements.ts
@@ -1,0 +1,4 @@
+import BesselianElementsDownloader from './BesselianElementsDownloader';
+
+const downloader = new BesselianElementsDownloader();
+downloader.run();