+++ src/earth/constants/diameters.ts
@@ -1,2 +1,0 @@
-// Source: https://nssdc.gsfc.nasa.gov/planetary/factsheet/
-export const DIAMETER_EARTH = 12756;