+++ src/solarEclipse/createSolarEclipseCircumstances.ts
@@ -1,0 +1,8 @@
+import SolarEclipseCircumstances from './SolarEclipseCircumstances';
+import {TimeLocalDependentCircumstances} from './types/circumstancesTypes';
+
+export default function createSolarEclipseCircumstances(
+    circumstances: TimeLocalDependentCircumstances,
+): SolarEclipseCircumstances {
+    return new SolarEclipseCircumstances(circumstances);
+}