+++ src/solarEclipse/index.ts
@@ -1,4 +1,4 @@
-import createSolarEclipse from './createSolarEclipse';
+import * as createSolarEclipse from './createSolarEclipse';
 import solarEclipseExists from './solarEclipseExists';
 
 export {createSolarEclipse, solarEclipseExists};